fastapi-master/requirements.txt
astroid>=2.3.3
Click>=7.0
colorama>=0.4.3
fastapi>=0.65.2
gunicorn>=20.0.4
h11>=0.9.0
isort>=4.3.21
lazy-object-proxy>=1.4.3
mccabe>=0.6.1
pydantic>=1.6.2
pylint>=2.4.4
six>=1.14.0
starlette>=0.12.9
uvicorn>=0.11.7
websockets>=9.1
wrapt>=1.11.2
uvloop>=0.14.0
Pillow>=9.2.0
numpy>=1.23.1
tensorflow>=2.9.1